DESeq2 Analysis
DESeq2 uses a statistical model to calculate the difference in gene expression between two or more groups of samples. It starts by estimating the variance of gene expression levels and then fits a negative binomial distribution to each gene. Here we use gene_count_matrix and set metadata & sample data for DESeq2 analysis. And then visualize using a Volcano plot and a heatmap.