RNA-Seq-DESeq2
DESeq2 is a tool popularly used for Differential gene analysis. Here we use it for RNA sequence. It takes gene read count data, accounts for the variability in sequencing, and uses statistical models, including shrinkage estimators for dispersion and fold change, to identify genes with significant expression changes between different experimental conditions. The following steps we use for DESeq2 are:- 
1. Normalization
2. Dispersion estimation
3. Differential expression analysis

We expect or generate a table of results, including the log2 fold change, p-value, and adjusted p-value (FDR) for each gene, making git easy and simpler to identify the most significant candidates.